+++ src/v0/destinations/fb_custom_audience/transform.js
@@ -1,6 +1,4 @@
-const sha256 = require('sha256');
 const get = require('get-value');
-
 const {
   defaultRequestConfig,
   defaultPostRequestConfig,
@@ -10,12 +8,11 @@ const {
   getSuccessRespEvents,
   getErrorRespEvents,
   returnArrayOfSubarrays,
-  isDefinedAndNotNull,
   flattenMap,
   handleRtTfSingleEventError,
   getDestinationExternalIDInfoForRetl,
 } = require('../../util');
-
+const { prepareDataField, getSchemaForEventMappedToDest, batchingWithPayloadSize } = require('./util');
 const {
   getEndPoint,
   schemaFields,
@@ -31,19 +28,6 @@ const {
   TransformationError,
   ConfigurationError,
 } = require('../../util/errorTypes');
-
-const getSchemaForEventMappedToDest = (message) => {
-  const mappedSchema = get(message, 'context.destinationFields');
-  if (!mappedSchema) {
-    throw new InstrumentationError(
-      'context.destinationFields is required property for events mapped to destination ',
-    );
-  }
-  // context.destinationFields has 2 possible values. An Array of fields or Comma seperated string with field names
-  let userSchema = Array.isArray(mappedSchema) ? mappedSchema : mappedSchema.split(',');
-  userSchema = userSchema.map((field) => field.trim());
-  return userSchema;
-};
 
 const responseBuilderSimple = (payload, audienceId) => {
   if (payload) {
@@ -64,136 +48,6 @@ const responseBuilderSimple = (payload, audienceId) => {
   // fail-safety for developer error
   throw new TransformationError(`Payload could not be constructed`);
 };
-
-// function responsible to ensure the user inputs are passed according to the allowed format
-
-const ensureApplicableFormat = (userProperty, userInformation) => {
-  let updatedProperty;
-  let userInformationTrimmed;
-  userInformation = userInformation.toString();
-  switch (userProperty) {
-    case 'EMAIL':
-      updatedProperty = userInformation.trim().toLowerCase();
-      break;
-    case 'PHONE':
-      // remove all non-numerical characters
-      updatedProperty = userInformation.replace(/\D/g, '');
-      // remove all leading zeros
-      updatedProperty = updatedProperty.replace(/^0+/g, '');
-      break;
-    case 'GEN':
-      updatedProperty =
-        userInformation.toLowerCase() === 'f' || userInformation.toLowerCase() === 'female'
-          ? 'f'
-          : 'm';
-      break;
-    case 'DOBY':
-      updatedProperty = userInformation.trim().replace(/\./g, '');
-      break;
-    case 'DOBM':
-      userInformationTrimmed = userInformation.replace(/\./g, '');
-      if (userInformationTrimmed.length < 2) {
-        updatedProperty = `0${userInformationTrimmed}`;
-      } else {
-        updatedProperty = userInformationTrimmed;
-      }
-      break;
-    case 'DOBD':
-      userInformationTrimmed = userInformation.replace(/\./g, '');
-      if (userInformationTrimmed.length < 2) {
-        updatedProperty = `0${userInformationTrimmed}`;
-      } else {
-        updatedProperty = userInformationTrimmed;
-      }
-      break;
-    case 'LN':
-    case 'FN':
-    case 'FI':
-      if (userProperty !== 'FI') {
-        updatedProperty = userInformation.toLowerCase().replace(/[!#$%&@A-Za-z]/g, '');
-      } else {
-        updatedProperty = userInformation.toLowerCase().replace(/[^!#$%&,.?@A-Za-z]/g, '');
-      }
-      break;
-    case 'MADID':
-      updatedProperty = userInformation.toLowerCase();
-      break;
-    case 'COUNTRY':
-      updatedProperty = userInformation.toLowerCase();
-      break;
-    case 'ZIP':
-      userInformationTrimmed = userInformation.replace(/\s/g, '');
-      updatedProperty = userInformationTrimmed.toLowerCase();
-      break;
-    case 'ST':
-    case 'CT':
-      updatedProperty = userInformation
-        .replace(/[^ A-Za-z]/g, '')
-        .replace(/\s/g, '')
-        .toLowerCase();
-      break;
-    case 'EXTERN_ID':
-      updatedProperty = userInformation;
-      break;
-    default:
-      throw new ConfigurationError(`The property ${userProperty} is not supported`);
-  }
-  return updatedProperty;
-};
-
-// Function responsible for making the data field without payload object
-// Based on the "isHashRequired" value hashing is explicitly enabled or disabled
-
-const prepareDataField = (
-  userSchema,
-  userUpdateList,
-  isHashRequired,
-  disableFormat,
-  skipVerify,
-) => {
-  const data = [];
-  let updatedProperty;
-  let dataElement;
-  userUpdateList.forEach((eachUser) => {
-    dataElement = [];
-    userSchema.forEach((eachProperty) => {
-      // if skip verify is true we replace undefined/null user properties with empty string
-      let userProperty = eachUser[eachProperty];
-      if (skipVerify && !isDefinedAndNotNull(userProperty)) {
-        userProperty = '';
-      }
-      if (isDefinedAndNotNull(userProperty)) {
-        if (isHashRequired) {
-          if (!disableFormat) {
-            // when user requires formatting
-            updatedProperty = ensureApplicableFormat(eachProperty, userProperty);
-          } else {
-            // when user requires hashing but does not require formatting
-            updatedProperty = userProperty;
-          }
-        } else {
-          // when hashing is not required
-          updatedProperty = userProperty;
-        }
-        if (isHashRequired && eachProperty !== 'MADID' && eachProperty !== 'EXTERN_ID') {
-          // for MOBILE_ADVERTISER_ID, MADID,EXTERN_ID hashing is not required ref: https://developers.facebook.com/docs/marketing-api/audiences/guides/custom-audiences#hash
-          updatedProperty = `${updatedProperty}`;
-          dataElement.push(sha256(updatedProperty));
-        } else {
-          dataElement.push(updatedProperty);
-        }
-      } else {
-        throw new ConfigurationError(
-          `Configured Schema field ${eachProperty} is missing in one or more user records`,
-        );
-      }
-    });
-    data.push(dataElement);
-  });
-
-  return data;
-};
-
 // Function responsible prepare the payload field of every event parameter
 
 const preparePayload = (
@@ -218,7 +72,7 @@ const preparePayload = (
     disableFormat,
     skipVerify,
   );
-  return prepareFinalPayload;
+  return batchingWithPayloadSize(prepareFinalPayload);
 };
 
 // Function responsible for building the parameters for each event calls
@@ -263,7 +117,7 @@ const prepareResponse = (
   if (Object.keys(dataSource).length > 0) {
     paramsPayload.data_source = dataSource;
   }
-  prepareParams.payload = preparePayload(
+  const payloadBatches = preparePayload(
     allowedAudienceArray,
     userSchema,
     paramsPayload,
@@ -271,16 +125,47 @@ const prepareResponse = (
     disableFormat,
     skipVerify,
   );
-
-  return prepareParams;
+  const respList = [];
+  payloadBatches.forEach((payloadBatch) => {
+    const response = {
+      ...prepareParams,
+      payload: payloadBatch,
+    };
+    respList.push(response);
+  });
+  return respList;
 };
 
-const processEvent = (message, destination) => {
-  let response;
-  const respList = [];
+/**
+ * Prepare to send events array 
+ * @param {*} message 
+ * @param {*} destination 
+ * @returns 
+ */
+const prepareToSendEvents = (message, destination, audienceChunksArray, userSchema,
+  isHashRequired, operation) => {
   const toSendEvents = [];
-  let wrappedResponse = {};
-
+  audienceChunksArray.forEach((allowedAudienceArray) => {
+    const responseArray = prepareResponse(
+      message,
+      destination,
+      allowedAudienceArray,
+      userSchema,
+      isHashRequired,
+    );
+    responseArray.forEach((response) => {
+      const wrappedResponse = {
+        responseField: response,
+        operationCategory: operation,
+      };
+      toSendEvents.push(wrappedResponse);
+    })
+  });
+  return toSendEvents;
+};
+const processEvent = (message, destination) => {
+  const respList = [];
+  let toSendEvents = [];
   let { userSchema } = destination.Config;
   const { isHashRequired, audienceId, maxUserCount } = destination.Config;
   if (!message.type) {
@@ -323,39 +208,13 @@ const processEvent = (message, destination) => {
   // when "remove" is present in the payload
   if (isDefinedAndNotNullAndNotEmpty(listData[USER_DELETE])) {
     const audienceChunksArray = returnArrayOfSubarrays(listData[USER_DELETE], maxUserCountNumber);
-    audienceChunksArray.forEach((allowedAudienceArray) => {
-      response = prepareResponse(
-        message,
-        destination,
-        allowedAudienceArray,
-        userSchema,
-        isHashRequired,
-      );
-      wrappedResponse = {
-        responseField: response,
-        operationCategory: USER_DELETE,
-      };
-      toSendEvents.push(wrappedResponse);
-    });
+    toSendEvents = prepareToSendEvents(message, destination, audienceChunksArray, userSchema, isHashRequired, USER_DELETE);
   }
 
   // When "add" is present in the payload
   if (isDefinedAndNotNullAndNotEmpty(listData[USER_ADD])) {
     const audienceChunksArray = returnArrayOfSubarrays(listData[USER_ADD], maxUserCountNumber);
-    audienceChunksArray.forEach((allowedAudienceArray) => {
-      response = prepareResponse(
-        message,
-        destination,
-        allowedAudienceArray,
-        userSchema,
-        isHashRequired,
-      );
-      wrappedResponse = {
-        responseField: response,
-        operationCategory: USER_ADD,
-      };
-      toSendEvents.push(wrappedResponse);
-    });
+    toSendEvents.push(...prepareToSendEvents(message, destination, audienceChunksArray, userSchema, isHashRequired, USER_ADD));
   }
   toSendEvents.forEach((sendEvent) => {
     respList.push(responseBuilderSimple(sendEvent, operationAudienceId));
